requirements.txt
aniso8601==9.0.1
api==0.0.7
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Flask==3.0.3
Flask-RESTful==0.3.10
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysql-connector-python==9.0.0
nose==1.3.7
pytz==2024.2
requests==2.32.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.0.4
